+++ src/main/java/org/kainos/ea/all_tables_apiApplication.java
@@ -5,7 +5,7 @@ import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
 import io.federecio.dropwizard.swagger.SwaggerBundle;
 import io.federecio.dropwizard.swagger.SwaggerBundleConfiguration;
-import org.kainos.ea.resources.JobRoleController;
+import org.kainos.ea.controller.JobRoleController;
 
 public class all_tables_apiApplication extends Application<all_tables_apiConfiguration> {
