+++ src/main/java/org/kainos/ea/core/JobRoleValidator.java
@@ -1,9 +1,9 @@
 package org.kainos.ea.core;
 
 
-import org.kainos.ea.cli.JobRoleRequest;
-import org.kainos.ea.client.JobRoleNameTooLongException;
-import org.kainos.ea.client.JobRoleSpecificationTooLongException;
+import org.kainos.ea.model.JobRoleRequest;
+import org.kainos.ea.exception.JobRoleNameTooLongException;
+import org.kainos.ea.exception.JobRoleSpecificationTooLongException;
 
 public class JobRoleValidator {
