+++ src/main/java/org/kainos/ea/exception/FailedToGetJobRoleException.java
@@ -1,4 +1,4 @@
-package org.kainos.ea.client;
+package org.kainos.ea.exception;
 
 public class FailedToGetJobRoleException extends Exception {
     @Override
