+++ src/main/java/org/kainos/ea/exception/JobRoleSpecificationTooLongException.java
@@ -1,4 +1,4 @@
-package org.kainos.ea.client;
+package org.kainos.ea.exception;
 
 public class JobRoleSpecificationTooLongException extends Throwable {
