+++ src/main/java/org/kainos/ea/model/JobRole.java
@@ -1,4 +1,4 @@
-package org.kainos.ea.cli;
+package org.kainos.ea.model;
 
 public class JobRole {
 
@@ -7,27 +7,17 @@ public class JobRole {
     private String name;
     private int bandId;
     private Band band;
+    private Capability capability;
 
     private String specification;
 
-    public JobRole(int id, String name, int bandId) {
-        this.setId(id);
-        this.setName(name);
-        this.setBandId(bandId);
-    }
-
-    public JobRole(int id, String name, Band band) {
+    public JobRole(int id, String name, Band band, Capability capability) {
         this.setId(id);
         this.setName(name);
         this.setBand(band);
+        this.setCapability(capability);
     }
 
-    public JobRole(int id, String name, int bandId, String specifictaion) {
-        setId(id);
-        setName(name);
-        setBandId(bandId);
-        setSpecification(specifictaion);
-    }
 
     public int getId() {
         return id;
@@ -67,5 +57,13 @@ public class JobRole {
 
     public void setBand(Band band) {
         this.band = band;
+    }
+
+    public Capability getCapability() {
+        return capability;
+    }
+
+    public void setCapability(Capability capability) {
+        this.capability = capability;
     }
 }
